backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gpxpy==1.5.0
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
contextily==1.4.0
geopandas==0.13.2
opencv-python-headless==4.8.1.78
scikit-image==0.21.0
rasterio==1.3.8
scipy==1.15.3